+++ R/drdid_imp_panel.R
@@ -14,7 +14,7 @@ NULL
 #' @param D An \eqn{n} x \eqn{1} vector of Group indicators (=1 if observation is treated in the post-treatment, =0 otherwise).
 #' @param covariates An \eqn{n} x \eqn{k} matrix of covariates to be used in the propensity score and regression estimation.
 #' If covariates = NULL, this leads to an unconditional DiD estimator.
-#' @param i.weights An \eqn{n} x \eqn{1} vector of weights to be used. If NULL, then every observation has the same weights.
+#' @param i.weights An \eqn{n} x \eqn{1} vector of weights to be used. If NULL, then every observation has the same weights. The weights are normalized and therefore enforced to have mean 1 across all observations.
 #' @param boot Logical argument to whether bootstrap should be used for inference. Default is FALSE.
 #' @param boot.type Type of bootstrap to be performed (not relevant if \code{boot = FALSE}). Options are "weighted" and "multiplier".
 #' If \code{boot = TRUE}, default is "weighted".
@@ -108,6 +108,9 @@ drdid_imp_panel <-function(y1, y0, D, covariates, i.weights = NULL, boot = FALSE
   if(is.null(i.weights)) {
     i.weights <- as.vector(rep(1, n))
   } else if(min(i.weights) < 0) stop("i.weights must be non-negative")
+
+  # Normalize weights
+  i.weights <- i.weights/mean(i.weights)
   #-----------------------------------------------------------------------------
   #Compute the Pscore using the pscore.cal
   pscore.ipt <- pscore.cal(D, int.cov, i.weights = i.weights, n = n)